refactor(trace): unify timestamps to u64 across record types
now_millis() previously returned u128, and the V1Alpha1 record types stored
timestamps as u128 even though every wire boundary (HTTP Api*Record types,
the InspectServerSession descriptor) clamped them to u64 via api_millis().

The clamp was harmless in practice (millis since epoch fits u64 until year
~292 million) but architecturally noisy: every From impl had a lossy cast,
every API surface drifted. Use u64 end to end.

- now_millis() -> u64
- *RecordV1Alpha1 timestamp fields: u128 -> u64
- AppProbe/AppAnalysis/AppDistillation/AppValidation::created_at_millis: u128 -> u64
- InspectServerSession::started_at_millis: u128 -> u64
- CaptureContract::captured_at_unix_ms: u128 -> u64
- ClipboardLock timeout comparison no longer needs `as u128`
- api_millis() helper deleted (identity now)
- Api*Record <-> *RecordV1Alpha1 From impls drop lossy casts

+++ src/app/mod.rs
@@ -70,7 +70,7 @@ pub struct AppIdentity {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct AppProbe {
   pub probe_version: String,
-  pub created_at_millis: u128,
+  pub created_at_millis: u64,
   pub project_root: PathBuf,
   pub output_dir: PathBuf,
   pub app: AppIdentity,
@@ -114,7 +114,7 @@ pub struct AppValidateOutput {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct AppDistillation {
   pub distill_version: String,
-  pub created_at_millis: u128,
+  pub created_at_millis: u64,
   pub source_analysis_path: PathBuf,
   pub app_identity: AppIdentity,
   pub candidates: Vec<AppDistilledCandidate>,
@@ -124,7 +124,7 @@ pub struct AppDistillation {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct AppValidation {
   pub validate_version: String,
-  pub created_at_millis: u128,
+  pub created_at_millis: u64,
   pub source_distillation_path: PathBuf,
   pub source_analysis_path: PathBuf,
   pub app_identity: AppIdentity,
@@ -165,7 +165,7 @@ pub struct AppValidatedCandidate {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct AppAnalysis {
   pub analysis_version: String,
-  pub created_at_millis: u128,
+  pub created_at_millis: u64,
   pub probe_path: PathBuf,
   pub app_identity: AppIdentity,
   pub window_context: AppWindowContext,
